requirements.txt
pytest
pyyaml
selenium==4.13.00
Selenium-Screenshot
Faker